requirements.txt
alembic==1.0.10
appdirs==1.4.3
attrs==19.1.0
backcall==0.1.0
black==19.3b0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
decorator==4.4.0
Faker==1.0.7
Flask==1.0.3
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
httpie==1.0.2
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-slugify==3.0.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.5
text-unidecode==1.2
toml==0.10.0
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
